Splitting _model.py step 2
- loader/types.py -- non-stac specific
- model.py -- stac specific
