+++ odc/stac/_mdtools.py
@@ -49,14 +49,13 @@ from pystac.extensions.projection import ProjectionExtension
 from pystac.extensions.raster import RasterBand, RasterExtension
 from toolz import dicttoolz
 
+from .loader.types import RasterBandMetadata, RasterSource
 from .model import (
     BandKey,
     BandQuery,
     MDParseConfig,
     ParsedItem,
-    RasterBandMetadata,
     RasterCollectionMetadata,
-    RasterSource,
 )
 
 T = TypeVar("T")
