+++ tests/test_reader.py
@@ -8,7 +8,7 @@ from numpy.testing import assert_array_equal
 from odc.geo.geobox import GeoBox
 from odc.geo.xr import xr_zeros
 
-from odc.stac._model import RasterLoadParams, RasterSource
+from odc.stac import RasterLoadParams, RasterSource
 from odc.stac._reader import (
     pick_overview,
     resolve_dst_dtype,
